backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
flask-migrate==4.0.7
flask-limiter==4.1.1
gunicorn==25.0.2
python-dotenv==1.2.1
requests==2.32.5
boto3==1.42.43
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
scikit-learn==1.8.0
numpy==2.4.1
scipy==1.17.0
beautifulsoup4==4.14.3
duckduckgo-search==6.3.7
pydantic==2.12.5
mangum==0.19.0
pyyaml==6.0.2

# ── LangGraph multi-agent system ───────────────────────────────────────────────
langgraph>=0.2.0
langchain-aws>=0.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.12.0
